Create README.md
This app is connected to a Firebase Realtime Database and allows users to both write/like endorsements for their work colleagues. The endorsements are then uploaded in real-time to Firebase and a user can delete same by clicking on it anywhere apart from the heart icon. A user can like an endorsement by clicking on the heart icon.